Fixed a lot of stuff!
Displaying multiple characters now work! (the value of rdx was modified by displayOneCharacter / dispFuncPtr, for some reason. I just saved it somewhere in saveStruct.rdx and restored it after the call)
The JIT compiler crashed after executing the compiled code, which was due to the same function (displayOneCharacter) messing up the stack, and thus the call stack. The fix, although kinda messy, was only to move the stack registers by 255 bytes during the execution of the cursed function and restoring rsp and rbp's values after the call had ended.
There's definitely something wrong with that function, but I can't understand it at all.

Also, all the comments are now in English, for consistency